+++ ada_verona/database/dataset/experiment_dataset.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 
 from typing_extensions import Self
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
+from ada_verona.database.dataset.data_point import DataPoint
 
 
 class ExperimentDataset(ABC):
