+++ ada_verona/database/dataset/image_file_dataset.py
@@ -6,8 +6,8 @@ import torch
 import torchvision.transforms as transforms
 from typing_extensions import Self
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
-from robustness_experiment_box.database.dataset.experiment_dataset import ExperimentDataset
+from ada_verona.database.dataset.data_point import DataPoint
+from ada_verona.database.dataset.experiment_dataset import ExperimentDataset
 
 
 @dataclass
