+++ ada_verona/database/verification_result.py
