+++ ada_verona/dataset_sampler/dataset_sampler.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
-from robustness_experiment_box.database.dataset.experiment_dataset import ExperimentDataset
-from robustness_experiment_box.database.network import Network
+from ada_verona.database.dataset.experiment_dataset import ExperimentDataset
+from ada_verona.database.network import Network
 
 
 class DatasetSampler(ABC):
