+++ ada_verona/epsilon_value_estimator/epsilon_value_estimator.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 
-from robustness_experiment_box.database.epsilon_value_result import EpsilonValueResult
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.verification_module.verification_module import VerificationModule
+from ada_verona.database.epsilon_value_result import EpsilonValueResult
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.verification_module.verification_module import VerificationModule
 
 
 class EpsilonValueEstimator(ABC):
