+++ ada_verona/verification_module/attack_estimation_module.py
@@ -3,13 +3,13 @@ import time
 import torch
 from autoverify.verifier.verification_result import CompleteVerificationData
 
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.database.verification_result import VerificationResult
-from robustness_experiment_box.verification_module.attacks.attack import Attack
-from robustness_experiment_box.verification_module.property_generator.one2any_property_generator import (
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.database.verification_result import VerificationResult
+from ada_verona.verification_module.attacks.attack import Attack
+from ada_verona.verification_module.property_generator.one2any_property_generator import (
     One2AnyPropertyGenerator,
 )
-from robustness_experiment_box.verification_module.verification_module import VerificationModule
+from ada_verona.verification_module.verification_module import VerificationModule
 
 
 class AttackEstimationModule(VerificationModule):
