+++ ada_verona/verification_module/attacks/attack.py
