+++ ada_verona/verification_module/attacks/auto_attack_wrapper.py
@@ -2,7 +2,7 @@ from autoattack import AutoAttack
 from torch import Tensor
 from torch.nn.modules import Module
 
-from robustness_experiment_box.verification_module.attacks.attack import Attack
+from ada_verona.verification_module.attacks.attack import Attack
 
 
 class AutoAttackWrapper(Attack):
