+++ ada_verona/verification_module/attacks/fgsm_attack.py
@@ -2,7 +2,7 @@ import torch
 from torch import Tensor, nn
 from torch.nn.modules import Module
 
-from robustness_experiment_box.verification_module.attacks.attack import Attack
+from ada_verona.verification_module.attacks.attack import Attack
 
 
 class FGSMAttack(Attack):
