+++ ada_verona/verification_module/test_verification_module.py
@@ -4,9 +4,9 @@ from autoverify.verifier.verification_result import CompleteVerificationData
 from result import Ok
 from torch import Tensor, load
 
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.database.verification_result import VerificationResult
-from robustness_experiment_box.verification_module.verification_module import VerificationModule
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.database.verification_result import VerificationResult
+from ada_verona.verification_module.verification_module import VerificationModule
 
 
 class TestVerificationModule(VerificationModule):
