+++ ada_verona/verification_module/verification_module.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 
 from autoverify.verifier.verification_result import CompleteVerificationData
 
-from robustness_experiment_box.database.verification_context import VerificationContext
+from ada_verona.database.verification_context import VerificationContext
 
 
 class VerificationModule(ABC):
