+++ tests/test_data_sampler/test_dataset_sampler.py
@@ -1,6 +1,6 @@
 import pytest
 
-from robustness_experiment_box.dataset_sampler.dataset_sampler import DatasetSampler
+from ada_verona.dataset_sampler.dataset_sampler import DatasetSampler
 
 
 def test_cannot_instantiate_datasetsampler():
