+++ tests/test_data_sampler/test_prediction_based_sampler.py
@@ -2,9 +2,9 @@ from pathlib import Path
 
 import pytest
 
-from robustness_experiment_box.database.dataset.image_file_dataset import ImageFileDataset
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.dataset_sampler.predictions_based_sampler import PredictionsBasedSampler
+from ada_verona.database.dataset.image_file_dataset import ImageFileDataset
+from ada_verona.database.network import Network
+from ada_verona.dataset_sampler.predictions_based_sampler import PredictionsBasedSampler
 
 
 @pytest.fixture
