+++ tests/test_database/conftest.py
@@ -2,12 +2,12 @@ import onnx
 import pytest
 import torch
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
-from robustness_experiment_box.database.epsilon_value_result import EpsilonValueResult
-from robustness_experiment_box.database.experiment_repository import ExperimentRepository
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.database.torch_model_wrapper import TorchModelWrapper
-from robustness_experiment_box.database.verification_context import VerificationContext
+from ada_verona.database.dataset.data_point import DataPoint
+from ada_verona.database.epsilon_value_result import EpsilonValueResult
+from ada_verona.database.experiment_repository import ExperimentRepository
+from ada_verona.database.network import Network
+from ada_verona.database.torch_model_wrapper import TorchModelWrapper
+from ada_verona.database.verification_context import VerificationContext
 
 
 class MockVerificationContext:
