+++ tests/test_database/test_dataset/test_pytorch_experiment_dataset.py
@@ -2,8 +2,8 @@ import pytest
 import torch
 from torch.utils.data import Dataset
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
-from robustness_experiment_box.database.dataset.pytorch_experiment_dataset import PytorchExperimentDataset
+from ada_verona.database.dataset.data_point import DataPoint
+from ada_verona.database.dataset.pytorch_experiment_dataset import PytorchExperimentDataset
 
 
 class MockDataset(Dataset):
