+++ tests/test_database/test_network.py
@@ -1,7 +1,7 @@
 import onnx
 
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.database.torch_model_wrapper import TorchModelWrapper
+from ada_verona.database.network import Network
+from ada_verona.database.torch_model_wrapper import TorchModelWrapper
 
 
 def test_network_initialization(network):
