+++ tests/test_database/test_verification_result.py
@@ -1,4 +1,4 @@
-from robustness_experiment_box.database.verification_result import VerificationResult
+from ada_verona.database.verification_result import VerificationResult
 
 
 def test_verification_result_values():
