+++ tests/test_epsilon_value_estimator/conftest.py
@@ -3,10 +3,10 @@ from pathlib import Path
 import pytest
 from autoverify.verifier.verification_result import CompleteVerificationData
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.verification_module.verification_module import VerificationModule
+from ada_verona.database.dataset.data_point import DataPoint
+from ada_verona.database.network import Network
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.verification_module.verification_module import VerificationModule
 
 
 @pytest.fixture
