+++ tests/test_verification_module/attacks/test_attack.py
@@ -1,6 +1,6 @@
 import pytest
 
-from robustness_experiment_box.verification_module.attacks.attack import Attack
+from ada_verona.verification_module.attacks.attack import Attack
 
 
 def test_cannot_instantiate_property_generator():
