+++ tests/test_verification_module/conftest.py
@@ -3,15 +3,15 @@ from autoverify.verifier.verification_result import CompleteVerificationData
 from result import Ok
 from torch import load
 
-from robustness_experiment_box.database.dataset.data_point import DataPoint
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.verification_module.attack_estimation_module import AttackEstimationModule
-from robustness_experiment_box.verification_module.auto_verify_module import AutoVerifyModule
-from robustness_experiment_box.verification_module.property_generator.one2any_property_generator import (
+from ada_verona.database.dataset.data_point import DataPoint
+from ada_verona.database.network import Network
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.verification_module.attack_estimation_module import AttackEstimationModule
+from ada_verona.verification_module.auto_verify_module import AutoVerifyModule
+from ada_verona.verification_module.property_generator.one2any_property_generator import (
     One2AnyPropertyGenerator,
 )
-from robustness_experiment_box.verification_module.test_verification_module import TestVerificationModule
+from ada_verona.verification_module.test_verification_module import TestVerificationModule
 
 
 @pytest.fixture
