+++ tests/test_verification_module/property_generator/test_property_generator.py
@@ -1,6 +1,6 @@
 import pytest
 
-from robustness_experiment_box.verification_module.property_generator.property_generator import PropertyGenerator
+from ada_verona.verification_module.property_generator.property_generator import PropertyGenerator
 
 
 def test_cannot_instantiate_property_generator():
