+++ tests/test_verification_module/test_auto_verify_module.py
@@ -4,15 +4,15 @@ import numpy as np
 import pytest
 from autoverify.verifier.verification_result import CompleteVerificationData
 
-from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.verification_module.auto_verify_module import (
+from ada_verona.database.verification_context import VerificationContext
+from ada_verona.verification_module.auto_verify_module import (
     parse_counter_example,
     parse_counter_example_label,
 )
-from robustness_experiment_box.verification_module.property_generator.one2any_property_generator import (
+from ada_verona.verification_module.property_generator.one2any_property_generator import (
     One2AnyPropertyGenerator,
 )
-from robustness_experiment_box.verification_module.property_generator.one2one_property_generator import (
+from ada_verona.verification_module.property_generator.one2one_property_generator import (
     One2OnePropertyGenerator,
 )
